[MOVE] Create GeneralModels and move files using general models
성별, 소셜 로그인 타입, 페이징 처리 모델과 응답 모델을 GeneralModels 폴더 내부에 두어
공통으로 처리되는 모델을 한 군데 모아두었습니다.
