+++ PLUB/Sources/Models/GeneralModels/Sex.swift
@@ -1,0 +1,17 @@
+//
+//  Sex.swift
+//  PLUB
+//
+//  Created by 홍승현 on 2023/04/05.
+//
+
+import Foundation
+
+/// 성별
+enum Sex: String, Codable {
+  /// 남성
+  case male = "M"
+  
+  /// 여성
+  case female = "F"
+}
